Fix engine tensor shapes, memory temp-store cleanup, and claim parsing
- forward.rs: fix broadcast shapes for eps/scale scalars, correct SiLU
  (was missing the up-projection multiply), flatten (b,seq,hidden) to
  2D for linear layers, and drop the stray batch dim before appending
  to the per-layer KV cache.
- storage.rs: back open_temp() with a real unique temp file instead of
  an empty path (redb requires a concrete path), and remove it on drop.
- logic.rs: extract_claim now takes the last word before '=' as the
  variable so qualifiers like "let"/"now" don't get folded into the
  variable name.
- Update CHANGELOG/TODO to reflect verified fmt/clippy/test/publish
  dry-run status and router latency benchmark results.

+++ crates/tpt-abyss-router/tests/router_tests.rs
@@ -1,4 +1,4 @@
-use tpt_abyss_router::{HeuristicRouter, RouterConfig, RouterConfigBuilder};
+use tpt_abyss_router::{HeuristicRouter, RouterConfigBuilder};
 use tpt_abyss_types::{LayerProgram, Position, TokenId};
 
 fn router() -> HeuristicRouter {
@@ -41,10 +41,10 @@ fn produced_program_is_valid() {
     for pos in 0..50u32 {
         let p = r
             .route_token(
-                TokenId(pos as u32 * 7),
+                TokenId(pos * 7),
                 Position(pos),
-                (pos as f32 / 50.0),
-                (pos as f32 / 50.0),
+                pos as f32 / 50.0,
+                pos as f32 / 50.0,
                 pos % 11 == 0,
             )
             .unwrap();
